flod_facilities_backend/requirements.txt
alembic==0.7.6
amqp==1.4.5
aniso8601==0.82
anyjson==0.3.3
billiard==3.3.0.17
blinker==1.3
boto==2.27.0
celery==3.1.11
coverage==3.7.1
Flask-Mail==0.9.0
Flask-RESTful==0.2.12
Flask-Script==0.6.7
Flask==0.10.1
GeoAlchemy2==0.2.4
httpretty==0.8.0
itsdangerous==0.24
Jinja2==2.7.2
kombu==3.0.15
Mako==0.9.1
MarkupSafe==0.21
Pillow==2.4.0
pip-tools==0.3.4
psycopg2==2.5.2
pycrypto==2.6.1
pytz==2014.2
requests==2.2.1
Shapely==1.2.17
six==1.6.1
SQLAlchemy==0.9.4
urllib3==1.8.2
Werkzeug==0.9.4
